fix:update db and get db data
